multi_agent_blog_writer/requirements.txt
# Basic requirements for the multi-agent blog writer
# Add any specific dependencies here if needed in the future
transformers
huggingface_hub
python-dotenv
streamlit
